+++ apps/web/utils/creatorProfile.ts
@@ -1,3 +1,6 @@
+import type { LoginUser } from "@/hooks/auth/useLogin";
+import { USER_STORAGE_KEY } from "@/utils/viewerProfile";
+
 export type Passwords = {
   current: string;
   next: string;
@@ -48,8 +51,97 @@ export const FORM_KEYS = {
 
 export type FormKeys = (typeof FORM_KEYS)[keyof typeof FORM_KEYS];
 
+export type CreatorBoot = Pick<ProfileForm, "firstName" | "lastName" | "email">;
+
+export const EMPTY_CREATOR_PROFILE_FORM: ProfileForm = {
+  firstName: "",
+  lastName: "",
+  company: "",
+  phone: "",
+  cvr: "",
+  address: "",
+  city: "",
+  postal: "",
+  reg: "",
+  account: "",
+  email: "",
+};
+
+export const EMPTY_CREATOR_BOOT: CreatorBoot = {
+  firstName: EMPTY_CREATOR_PROFILE_FORM.firstName,
+  lastName: EMPTY_CREATOR_PROFILE_FORM.lastName,
+  email: EMPTY_CREATOR_PROFILE_FORM.email,
+};
+
+const str = (value: unknown) => (typeof value === "string" ? value.trim() : "");
+
+export const getAvatarUrl = (avatarUrl?: string | null): string | null => {
+  if (typeof avatarUrl !== "string") return null;
+  return avatarUrl.length > 0 ? avatarUrl : null;
+};
+
+export const toOptionalString = (value: string): string | undefined =>
+  value || undefined;
+
+export function parseCreatorNameFromFullName(
+  fullName: string,
+): Pick<ProfileForm, "firstName" | "lastName"> {
+  const parts = fullName.split(/\s+/).filter(Boolean);
+  return {
+    firstName: parts[0] ?? EMPTY_CREATOR_BOOT.firstName,
+    lastName: parts.slice(1).join(" "),
+  };
+}
+
+export function readCreatorBoot(): CreatorBoot {
+  if (typeof window === "undefined") {
+    return EMPTY_CREATOR_BOOT;
+  }
+
+  try {
+    const raw = window.localStorage.getItem(USER_STORAGE_KEY);
+    const user = raw ? (JSON.parse(raw) as LoginUser) : null;
+    const email = str(user?.email);
+
+    const storedFirst = str(user?.firstName);
+    const storedLast = str(user?.lastName);
+    if (storedFirst || storedLast) {
+      return {
+        ...EMPTY_CREATOR_BOOT,
+        firstName: storedFirst,
+        lastName: storedLast,
+        email,
+      };
+    }
+
+    const fullName = str(user?.fullName);
+    if (fullName) {
+      return {
+        ...EMPTY_CREATOR_BOOT,
+        ...parseCreatorNameFromFullName(fullName),
+        email,
+      };
+    }
+
+    return { ...EMPTY_CREATOR_BOOT, email };
+  } catch {
+    return EMPTY_CREATOR_BOOT;
+  }
+}
+
+export function displayCreatorName(
+  form: Pick<ProfileForm, "firstName" | "lastName">,
+): string {
+  return [form.firstName, form.lastName]
+    .map((part) => part.trim())
+    .filter(Boolean)
+    .join(" ");
+}
+
 const CREATOR_PROFILE_UTILS = {
   FORM_KEYS,
+  EMPTY_CREATOR_PROFILE_FORM,
+  EMPTY_CREATOR_BOOT,
 };
 
 export default CREATOR_PROFILE_UTILS;
